Update how item polymorphism is handled.
Originally we had the item type id set on the items, and this would work
fine for reading an item that was written through this model, but it did
not work reading items directly from the postman API as the type info
was missing. To address this we have added an Item deserializer that
will deserialize items based on the presence or absence of properties on
the item nodes being read.